requirements.txt
# Core ML/Deep Learning
torch>=2.0.0
tensorflow>=2.15.0
tensorflow-recommenders>=0.7.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
polars>=0.20.0
scikit-learn>=1.3.0

# Image Processing & Embeddings
Pillow>=10.0.0
transformers>=4.35.0
open-clip-torch>=2.24.0
fashion-clip>=0.2.0

# Vector Database
pgvector>=0.2.0
psycopg2-binary>=2.9.0
faiss-cpu>=1.7.4

# API & Serving
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
redis>=5.0.0

# Background Removal & Image Processing
rembg>=2.0.50

# OpenAI for Listing Generation
openai>=1.10.0

# Experiment Tracking
mlflow>=2.10.0

# A/B Testing
growthbook>=1.0.0

# Arabic NLP
camel-tools>=1.5.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
httpx>=0.26.0
aiohttp>=3.9.0
loguru>=0.7.0
pyyaml>=6.0.0
typer>=0.9.0
rich>=13.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
